Expose FMC language detection as custom data endpoint
