+++ webscrambles/src/main/kotlin/org/worldcubeassociation/tnoodle/server/webscrambles/zip/folder/PrintingFolder.kt
@@ -11,16 +11,14 @@ import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.ScrambleDrawing
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.WCIFDataBuilder
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.WCIFDataBuilder.getCachedPdf
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.ActivityCode
-import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.Competition
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.Schedule
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.ScrambleSet
 import org.worldcubeassociation.tnoodle.server.webscrambles.wcif.model.extension.FmcLanguagesExtension
 import org.worldcubeassociation.tnoodle.server.webscrambles.zip.folder
 import org.worldcubeassociation.tnoodle.server.webscrambles.zip.model.Folder
 import java.time.LocalDate
-import java.util.*
 
-data class PrintingFolder(val uniqueTitles: Map<String, ScrambleDrawingData>, val globalTitle: String, val wcif: Competition) {
+data class PrintingFolder(val uniqueTitles: Map<String, ScrambleDrawingData>, val globalTitle: String, val wcifSchedule: Schedule) {
     val scrambleSheetsFlat = uniqueTitles.values.toList()
     val scrambleDrawingData = CompetitionDrawingData(globalTitle, scrambleSheetsFlat)
 
@@ -54,7 +52,6 @@ data class PrintingFolder(val uniqueTitles: Map<String, ScrambleDrawingData>, va
                         folder("Translations") {
                             val requestedTranslations = req.scrambleSet.findExtension<FmcLanguagesExtension>()
                                 ?.languageTags?.takeUnless { it.isEmpty() }
-                                ?: compileCompetitorLanguages().takeUnless { it.isEmpty() }
                                 ?: Translate.TRANSLATED_LOCALES.map { it.toLanguageTag() }
 
                             for (locale in Translate.TRANSLATED_LOCALES) {
@@ -78,9 +75,9 @@ data class PrintingFolder(val uniqueTitles: Map<String, ScrambleDrawingData>, va
                 }
             }
 
-            if (wcif.schedule.isNotEmpty()) {
+            if (wcifSchedule.isNotEmpty()) {
                 val orderedScramblesFolder = OrderedScramblesFolder(globalTitle, scrambleDrawingData)
-                val orderedScramblesNode = orderedScramblesFolder.assemble(wcif.schedule, generationDate, versionTag)
+                val orderedScramblesNode = orderedScramblesFolder.assemble(wcifSchedule, generationDate, versionTag)
 
                 folder(orderedScramblesNode)
             }
@@ -93,36 +90,7 @@ data class PrintingFolder(val uniqueTitles: Map<String, ScrambleDrawingData>, va
         }
     }
 
-    private fun compileCompetitorLanguages(): List<String> {
-        val competitorCountries = this.wcif.persons
-            .map { it.countryIso2 }
-            .map { it.isoString }
-            .distinct()
-
-        val countryLocales = competitorCountries.flatMap { getTranslatedCountryLanguage(it) }
-            .toSet() + Translate.DEFAULT_LOCALE
-
-        return countryLocales.distinct().map { it.toLanguageTag() }
-    }
-
     companion object {
         private fun Schedule.isNotEmpty() = leafActivities.isNotEmpty()
-
-        private val COUNTRY_ISO2_LOCALES = Locale.getAvailableLocales()
-            .groupBy { it.country }
-
-        private val TRANSLATED_LANGUAGE_ISO2 = Translate.TRANSLATED_LOCALES
-            .map { it.language }.toSet()
-
-        private fun getTranslatedCountryLanguage(iso2: String): List<Locale> {
-            val candidateLocales = COUNTRY_ISO2_LOCALES[iso2] ?: return emptyList()
-
-            val translatedLocales = candidateLocales.filter { it.language in TRANSLATED_LANGUAGE_ISO2 }
-
-            val localLanguages = translatedLocales.filter { Locale(it.language, iso2) in Translate.TRANSLATED_LOCALES }
-            val generalLanguages = translatedLocales.flatMap { Translate.TRANSLATED_LOCALES.filter { l -> l.language == it.language } }
-
-            return localLanguages.takeIf { it.isNotEmpty() } ?: generalLanguages
-        }
     }
 }
